Updating threshold and changing ignoreDev default
So this is taking an opposite approach for the threshold. If you
specify the threshold as moderate you will get moderate, high and
critical advisories.

Aslo changed the ignoreDev default from false to true. This is so
issues like the one earlier gets flagged.

Added the action column to show which root modules are affected.